feat: expose a cache cleanup function for unit tests
* move resource selectors to a separate file
